backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
browser-cookie3==0.20.1
playwright==1.54.0
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
jieba==0.42.1
